requirements.txt
APScheduler==3.8.1
boto3==1.19.6
Django==3.2.9
python-cei-crawler==0.1.2
django-environ==0.8.1
psycopg2==2.8.5
psycopg2-binary==2.8.5